backend/requirements.txt
alembic==1.15.2
asyncpg==0.30.0
bcrypt==5.0.0
cryptography==46.0.6
email-validator==2.2.0
fastapi==0.116.1
greenlet==3.2.4
httpx==0.28.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==6.4.0
sqlalchemy==2.0.43
uvicorn==0.35.0
websockets==15.0.1
